Module Sc_C.Cov_label

Coverage labels management

type 'a t = 'a Types.cov_label

type simple = [ `simple ] t

type hyper = [ `hyper ] t

type nonrec any = Types.any t

val id : _ t -> int

Returns the unique identifier of the given label.

val status : _ t -> Types.cov_label_status

Returns the status of the label.

val emitter : _ t -> string option

Returns the name of the tool who emited a status.

val location : _ t -> Types.location

Returns the location in the labelized file.

val is_covered : _ t -> bool

Returns true if the label status is 'Covered _'

val is_uncoverable : _ t -> bool

Returns true if the label status is 'Uncoverable'

val is_unknown : _ t -> bool

Returns true if the label status is 'Unknwon'

val max_label_id : _ t list -> int

Checks that the labels ids are contiguous and start at 1, then returns the max label id (which must be the total number of label). Raises Invalid_argument if it does not satisfy the specification.

Casting

As their name/type say.

val as_any : _ t -> any

val inspect : _ t -> [ `hyper | `simple ] t

Status

val show_status : Types.cov_label_status -> string

The string representation of a status.

val set_status : Types.cov_label_status -> 'a t -> 'a t

Replaces the label status by the one in argument.

val print_status : Types.cov_label_status Basics.PPrt.pp

Prints the string representation of a status.