Module Sc_C.Defs

Helpers for Cil datatypes manipulation.

Variable declarations

val var_name : Types.var -> string

val var_names : Types.vars -> string list

val varset : Types.vars -> Basics.Strings.t

varset vars is the set of all variables names in vars

val typ : Types.var -> Cil.typ

Attributes

val carray_type_attributes : Cil.attributes -> Types.carray_type_attrs

Returns details on the attributes of an array type.

val as_pointer_to_carray : ?size:[ `Int of int | `Var of string ] -> Types.var -> Types.var

val as_pointer_to_cstring : Types.var -> Types.var

val as_pointer_field_to_carray_field : ?size:[ `Int of int | `Var of string ] -> Cil.fieldinfo -> Cil.fieldinfo

val as_pointer_field_to_cstring_field : Cil.fieldinfo -> Cil.fieldinfo

val pointer_validity : Cil.attributes -> Types.pointer_validity

pointer_validity attrs retrieves the kind of a pointer variable or field from its attributes attrs: it is either a regular C array (the default when no specific attrribute is given), or a C string (null-terminated).

val const_attribute : Cil.attributes -> bool

val const_typ : Cil.typ -> bool

Accessors and predicates

exception ARG_INFO_ERROR of Cil.typ

val fun_arg_names : string -> Cil.file -> string list

fun_arg_names f file returns the names of f's formal arguments.

val fun_arg_typs : string -> Cil.file -> Cil.typ list

fun_arg_typs f file returns the type of f's formal arguments.

val fun_arg_unrolled_typs : string -> Cil.file -> Cil.typ list

fun_arg_unrolled_typs f file same as fun_arg_typs, but unrolls types.

val fun_args : string -> Cil.file -> Types.vars

fun_args f file returns the name and type of f's formal arguments.

val fun_unrolled_args : string -> Cil.file -> Types.vars

fun_unrolled_args f file same as fun_args, but unrolls types.

val fun_return_typ : string -> Cil.file -> Cil.typ

fun_return_typ f file returns f's return type.

val glob_types : Cil.file -> Cil.typ list

glob_types file Returns the type of the global variables

val glob_unrolled_types : Cil.file -> Cil.typ list

glob_unrolled_types file Same as glob_types but unrolls types

val glob_names : Cil.file -> string list

glob_types file Returns the name of the global variables

val glob_infos : Cil.file -> Types.vars

glob_infos file Returns the type and the name of the global variables

val glob_unrolled_infos : Cil.file -> Types.vars

glob_unrolled_infos file Same as glob_infos but unrolls types

val fold_typ_defs : Cil.file -> (Cil.global -> 'a -> 'a) -> 'a -> 'a

fold_typ_defs file f acc folds f over every type definition in file.

val diff_globals : Cil.file -> Cil.file -> Cil.file

diff_globals cil1 cil2 Performs a diff between cil1 and cil2 by relying on source file locations to check equality.

val replace_globals : f:(Cil.global -> Cil.global) -> Cil.file -> unit

visit_globals ~f cil replaces every global definition g of the given CIL file f with the result of f g.

val multidim_array_sizes : Cil.typ -> Types.carray_type_attrs list

Returns the list of subarrays sizes of the array type in argument. If the type is not an array, returns

val func_env : ?for_func:string -> Cil.file -> Types.func_env

Initializes the global variables of a file. For now, it ignores the for_func parameter.

val map_func_inputs : f:(Types.input_kind -> Types.var -> Types.var) -> Types.func_repr -> Types.func_repr

val cil_func : Cil.file -> string -> Types.func_repr

cil_func file f

Builds the simplified representation of the function f defined in file file.

val func_inputs : Types.func_repr -> Types.vars

val parse_file : [ `C ] Sc_sys.File.file -> Cil.file