Module Sc_C.Printer

Helpers for Cil datatypes manipulation.

Variable declarations

val pp_vars : Types.vars Fmt.t

val pp_formal_decls : Types.vars Fmt.t

val pp_global_decls : Types.vars Fmt.t

Access path

val pp_access_path : Types.access_path Fmt.t

val pp_typed_assignment : (Cil.typ * string * string) Fmt.t

pp_typed_assignment fmt (typ, dest, src) uses the formatter fmt to print an assignment from src to dest, the latter being of type typ.

Other stuff for printing and representing functions

val string_of_typ : ?enable_static_attr:bool -> Cil.typ -> string

string_of_typ ty Returns the string representation of a type.

val pp_exp : Cil.exp Fmt.t

Pretty prints an expression.

val pp_typ : Cil.typ Fmt.t

Pretty prints a type.

val pp_typ_defs : Cil.file Fmt.t

Pretty-prints every type definition from a given Cil file.

val pp_func_env : Types.func_env Fmt.t

Pretty prints an environment.

val pp_fundecl : ?enable_static_attr:bool -> Types.func_repr Fmt.t

Pretty prints a function representant as an extern C function declaration.

Pointer specifications

val pp_pointer_ref : Types.pointer_ref Fmt.t

val pp_pointer_constraint : Types.pointer_constraint Fmt.t

val emit_testcall : oracle_assessment:string -> emit_effective_inputs:'a Fmt.t -> entrypoint:Types.func_repr -> ?init_func:Types.func_repr -> ?oracle_func:Types.func_repr -> 'a -> Stdlib.Format.formatter -> unit

emit_testcall ~oracle_assessment ~emit_effective_inputs ~entrypoint ~init_func ~oracle_func effective_inputs ppf emits C code that encapsulates a call to the provided entrypoint function using the optional functions for initialization init_func and test oracle oracle_func.

If provided, init_func should not accept any argument; any result it returns is ignored.

If provided, oracle_func should accept the same list of arguments as entrypoint, plus an additional parameter with the return type of entrypoint (if not void). The oracle should also return an integral result, which is checked as non-null using a prediacte named oracle_assessment.

Errors

val pp_error : Types.error Fmt.t