Module Sc_C.Ptr_specs

val pointer_ref_of_string : string -> (Types.pointer_ref, Types.error) Stdlib.result

pointer_ref_of_string s parses the string s for a pointer reference. Expected format is <variable_name> | "{struct" <struct_name> "}" "." <field_name>.

val pointer_constraint_of_string : string -> (Types.pointer_constraint, Types.error) Stdlib.result

pointer_ref_of_string s parses the string s for a pointer reference. Expected format is <variable_name> | "{struct" <struct_name> "}" "." <field_name>.

val var_mem : pointer_var:string -> Types.pointer_refs -> bool

var_mem v gap_list holds whenever the size of an array referenced by a pointer variable v is constrained according to gap_list. .

val field_mem : struct_name:string -> pointer_field_name:string -> Types.pointer_refs -> bool

field_mem ~struct_name ~field_name gap_list returns true iff a field of a structure struct_name holds the size of the array referenced by a pointer field field_name in the same structure.

val find_size_var : pointer_var:string -> Types.pointer_constraints -> string option

find_size_var ~varname gap_list returns the name of the variable that holds the size of the array referenced by a pointer varname, if any.

val find_size_field : struct_name:string -> pointer_field_name:string -> Types.pointer_constraints -> string option

find_size_field ~struct_name ~field_name gap_list returns the name of the field of any structure struct_name that holds the size of the array referenced by a pointer field field_name in the same structure.