Module Sc_C.Types

Helpers for Cil datatypes manipulation.

Variable declarations

type var = Cil.typ * string * Cil.attributes

type vars = var list

Access path

type access_path_operation =

| Field_access of string
(*
access to a field given by its name
*)

| Index_access of int
(*
access to an array cell given by its index
*)

A concrete access path is built out of simple access operations

type access_path_suffix = access_path_operation NEL.t

An access path suffix is a non-empty list of access path operations

type access_path =

| Access_path_origin of string

| Access_path_suffix of access_path_suffix

| Access_path of {

origin : string;

suffix : access_path_suffix;

}

A concrete access path combines an origin variable name and/or a suffix

Attributes

type carray_type_attrs =

| Static_array of int

| Dynamic_array

C array type attributes.

Array attributes give its length and its static/dynamic characteristics.

Pointer validity

type pointer_validity = [

| `Carray_with_bound_length of int

| `Carray_with_length_field of string

| `Carray_with_unknown_length

| `Cstring

]

Functions

type func_env = {

glob_vars : vars;

}

The environment of a function. For now, only gathers the file-global variables.

type func_repr = {

func_name : string;

func_env : func_env;

func_rtyp : Cil.typ;

func_args : vars;

}

The simplified representation of a Cil function.

type input_kind =

| Formal

| Global

Coverage labels

type cov_label_view = {

cov_label_id : int;
(*
The unique ID of the proof objective
*)

cov_label_status : cov_label_status;
(*
The current status of the proof objective
*)

cov_label_orig_loc : location;
(*
The location of the proof objective in the original file
*)

cov_label_loc : location;
(*
The location in the labelized file
*)

cov_label_emitter : string option;
(*
The tool that emited the status
*)

cov_label_exec_time : float;
(*
Exec time of the proof
*)

}

and location = {

loc_file : string;

loc_line : int;

}

A location is a line in a file.

and cov_label_status =

| Unknown

| Covered of string list

| Uncoverable

Status of a coverage label / proof objective

type simple = [

| `simple

]

type hyper = [

| `hyper

]

type any = [

| simple

| hyper

| `any

]

type _ cov_label =

| S : cov_label_view -> [> simple ] cov_label

| H : cov_label_view -> [> hyper ] cov_label

| Any : [ simple | hyper ] cov_label -> [> `any ] cov_label

type pointer_refs = pointer_ref list

Representation for sets of pointer references

and pointer_ref =

| Variable of {

pointer_var : string;

}
(*
Reference to a pointer variable
*)

| Struct_field of {

struct_name : string;

pointer_field_name : string;

}
(*
Reference to a pointer field
*)

Type of pointer references

type pointer_constraints = pointer_constraint list

Representation for pointer constraints

and pointer_constraint =

| Distinct_variables of {

pointer_var : string;
(*
Name of the constrained pointer variable
*)

size_var : string;
(*
Name of the constraining size variable
*)

}
(*
Constraint between two distinct variables
*)

| From_same_struct of {

struct_name : string;
(*
The structure type name
*)

pointer_field_name : string;
(*
Name of the constrained pointer field
*)

size_field_name : string;
(*
Name of the constraining size field
*)

}
(*
Constraint between two fields of the same structure
*)

A single pointer constraint relates distinct variables or fields in the same structure.

Errors and exceptions

type error =

| Syntax_error of {

expected : parsed_item;

string : string;

}

and parsed_item =

| Pointer_reference

| Pointer_constraint

exception Unknown_function of string

exception Invalid_attribute_payload of Cil.attribute

exception Incompatible_attributes of string list

exception Missing_attribute of {

name : string;

reason : string;

}

exception UNSUPPORTED_TYPE of Ctypes_static.boxed_typ